% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreePhylip.R
\name{create.custom.species.tree}
\alias{create.custom.species.tree}
\title{Species Pruner}
\usage{
create.custom.species.tree(
  gene.list,
  suffix = "ordered_nuc_trimmed.fas",
  input.directory,
  tree.output.directory,
 
    ensembl_tree_version = "https://raw.githubusercontent.com/Ensembl/ensembl-compara/release/\\n                                       97/scripts/pipeline/species_tree.vertebrates.branch_len.nw"
)
}
\arguments{
\item{gene.list}{character; A list of multiple sequence alignments names}

\item{suffix}{character; Suffix/filetype of the files containing the multiple sequence
alignments}

\item{input.directory}{character; Where alignments are stored}

\item{tree.output.directory}{character; Where trees are going to be written}

\item{ensembl_tree_version}{character; Compara Ensembl newick species tree github site}
}
\value{
A txt file containing trees and a phylip-like fasta file.
}
\description{
This function allows you to prune the general species tree according to the
species present in your alignment. It also generates a phylip-like file for
codeml input.
}
\keyword{trees}
