% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblDownload.R
\name{ensembl.orthologue.download}
\alias{ensembl.orthologue.download}
\title{Download Ensembl Orthologues}
\usage{
ensembl.orthologue.download(gene.list, target_taxon, download.directory)
}
\arguments{
\item{gene.list}{A list of gene symbol names}

\item{target_taxon}{The NCBI number of the species you'd like to download}

\item{download.directory}{The directory where you'd like the downloaded multiple sequences to be stored}
}
\value{
A txt file containing sequences and species names in fasta format.
}
\description{
This function allows you to download 1-to-1 orthologues from Ensembl, given a target taxon.
}
\examples{
\dontrun{ensembl.orthologue.download(gene.list = c("RBFOX1", "FOXP2"), target_taxon = "9798", download.directory = ".")}
}
\keyword{download}
