% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{export.fasta.out}
\alias{export.fasta.out}
\title{Export info from list as fasta}
\usage{
export.fasta.out(list, output.directory, no.sequences = TRUE)
}
\arguments{
\item{list}{list; list obtained from download.orthologues be used as input}

\item{output.directory}{character; directory to write a fasta file for each
element in the list}

\item{no.sequences}{logical; if \code{TRUE}, returns a txt file with a list of genes
without sequences (Empty or No homologies)}
}
\value{
A fasta file
}
\description{
Utility to export the sequences as
a fasta file. It also attempts to create a file on your
desktop for each element of the gene list as  \code{*.fasta}.
}
\examples{
\dontrun{
export.fasta.out(ex.genes, "/mydirectory/", no.sequences = FALSE)}

}
