% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagTrees.R
\name{extract.tag.tree}
\alias{extract.tag.tree}
\title{Tag trees}
\usage{
extract.tag.tree(
  gene.list,
  m0.directory,
  classifier.directory,
  Ha.directory,
  Ho.directory
)
}
\arguments{
\item{gene.list}{character; A list of multiple sequence alignments names}

\item{m0.directory}{character; The directory where your PAML m0 files are located}

\item{classifier.directory}{Tcharacter; he directory where csv files indicating which species are foreground and background.}

\item{Ha.directory}{character; The directory where the alternative hypotesis of your PAML will be run}

\item{Ho.directory}{character; The directory where the null hypotesis of your PAML will be run}
}
\value{
A txt file in the Ha and Ho directories containing the trees tagged with #1 in the foreground lineage.
}
\description{
This function allows you to tag trees in foreground or background lineages.
}
\keyword{tag.trees}
