% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastClassify.R
\name{fast.classify}
\alias{fast.classify}
\title{Fast Foreground Classifier Function}
\usage{
fast.classify(gene.list, suffix, foreground.species)
}
\arguments{
\item{gene.list}{A list of multiple sequence alignments names}

\item{suffix}{Suffix/filetype of the files containing the multiple sequence alignments}

\item{foreground.species}{List of foreground species to be searched in the multiple sequences}
}
\value{
A csv file containing species names and classification into foreground or background lineages. Useful for the \link[selectionr]{extract.tag.tree}.
}
\description{
This function allows you to classify which of your sequences will be tested as foreground lineages.
}
\details{
This function makes use of a list of species that you consider to be foreground and outputs a csv for each gene you are testing.
}
\examples{
\dontrun{fast.classify(gene.list = c("RBFOX1", "FOXP2"), suffix = ".fasta", foreground.species = "homo_sapiens")}
}
\seealso{
\link[selectionr]{extract.tag.tree}
}
\keyword{foreground}
