% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_select.coxph.r
\name{forward_select.coxph}
\alias{forward_select.coxph}
\title{Forward Selection for Cox-PH}
\usage{
forward_select.coxph(
  formula,
  data,
  test,
  sort = NULL,
  rev = FALSE,
  max_ints = 2,
  complete_cases = FALSE,
  main_fx = TRUE,
  ...
)
}
\arguments{
\item{formula}{The formula for the starting point model}

\item{data}{The dataset to use}

\item{test}{A vector of test variables}

\item{sort}{Sort output by this variable}

\item{rev}{Reverse output}

\item{max_ints}{Maximum level of interactions}

\item{complete_cases}{Only include complete cases of data}

\item{main_fx}{Require all main effects in order to consider interactions. Default TRUE}

\item{...}{...}
}
\value{
An fs object
}
\description{
Forward Selection for Cox-PH
}
