% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_select.R
\name{forward_select}
\alias{forward_select}
\title{Wrapper for Select Function}
\usage{
forward_select(
  formula,
  data,
  test,
  method = NULL,
  family = NULL,
  sort = NULL,
  rev = FALSE,
  max_ints = 2,
  complete_cases = FALSE,
  main_fx = TRUE,
  ...
)
}
\arguments{
\item{formula}{The formula for the starting point model}

\item{data}{The dataset to use}

\item{test}{A vector of test variables}

\item{method}{Method to use}

\item{family}{Not used}

\item{sort}{Sort output by this variable}

\item{rev}{Reverse output}

\item{max_ints}{Maximum level of interactions}

\item{complete_cases}{Only include complete cases of data}

\item{main_fx}{Require all main effects in order to consider interactions. Default TRUE}

\item{...}{...}
}
\value{
An fs object
}
\description{
Wrapper for Select Function
}
