% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicLocations.R
\name{genomic.locator}
\alias{genomic.locator}
\title{Get genomic location from Ensembl Protein}
\usage{
genomic.locator(
  server = "https://jul2019.rest.ensembl.org",
  ensembl.protein.ID,
  original.number
)
}
\arguments{
\item{server}{The REST Ensembl version you'd like to access}

\item{ensembl.protein.ID}{Ensembl Protein ID}

\item{original.number}{Original position to query}
}
\value{
A data frame with gene name, start and end in human genomic positions.
}
\description{
This function allows you to get the geneomic location
 from aminoacid sites obtained from Ensembl proteins.
}
