% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interactions.r
\name{get_interactions}
\alias{get_interactions}
\title{Get interactions from Variables}
\usage{
get_interactions(rhs, max_ints = 2)
}
\arguments{
\item{rhs}{variables}

\item{max_ints}{max interaction level}
}
\description{
Get interactions from Variables
}
