% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmmCleaner.R
\name{HmmClean}
\alias{HmmClean}
\title{HmmClean}
\usage{
HmmClean(
  gene.list,
  suffix_AA = "_pranked.best.pep.fas",
  suffix_NT = "_pranked.best.nuc.fas",
  input.align.directory,
  output.clean.directory
)
}
\arguments{
\item{gene.list}{A gene list or file names}

\item{suffix_AA}{Suffix from the file containing the MSA in AA}

\item{suffix_NT}{Suffix from the file containing the MSA in NT}

\item{input.align.directory}{Directory where the alignments are located}

\item{output.clean.directory}{Directory where cleaned files will be written}
}
\value{
A fasta file and other files containing info about cleaned and masked NT
}
\description{
Calls HMMCleaner and cleans a fasta file (in development)
}
