% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginalPosition.R
\name{positions.provean}
\alias{positions.provean}
\title{Original sequence positions}
\usage{
positions.provean(
  gene.name,
 
    input.directory = "/Users/Usuario/Desktop/selection/results extensive/results omm_macse_extensive/2. Align omm_macse_extensive/",
  suffix = "_final_align_NT.aln",
  output.directory,
 
    classifier.directory = "/Users/Usuario/Desktop/selection/results extensive/results omm_macse_extensive/5. Classify omm_macse_extensive/",
 
    beb.sites.directory = "/Users/Usuario/Desktop/selection/results extensive/results omm_macse_extensive/9.Summary omm_macse_extensive/"
)
}
\arguments{
\item{gene.name}{A list of multiple sequence alignments names}

\item{input.directory}{Directory name where MSA is}

\item{suffix}{Suffix/filetype of the files containing the multiple sequence alignments}

\item{output.directory}{Where to output results}

\item{classifier.directory}{Where are species classified into foreground and background}

\item{beb.sites.directory}{Where are BEB sites stored}
}
\value{
A data frame with equivalences between masked and unmasked nucleotides
}
\description{
This function allows you to get gene original sequence, without gaps and without masked nucleotides.
}
\details{
This function makes use of a list of species that you consider to be foreground and outputs a csv for each gene you are testing.
}
\keyword{foreground}
