% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrankAligner.R
\name{prank.align}
\alias{prank.align}
\title{Prank Aligner Wrapper}
\usage{
prank.align(
  gene.list,
  input.file.directory,
  output.file.directory,
  software.directory
)
}
\arguments{
\item{gene.list}{A list of multiple sequence alignments names}

\item{input.file.directory}{Suffix/filetype of the files containing the multiple sequence alignments}

\item{output.file.directory}{List of foreground species to be searched in the multiple sequences}

\item{software.directory}{Software directory}
}
\value{
An aligned fasta.
}
\description{
This function allows you to align sequences with prank aligner.
}
\keyword{aligner}
