% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fs.r
\name{print.fs}
\alias{print.fs}
\title{S3 method for print for fs class}
\usage{
\method{print}{fs}(x, digits = 3, ...)
}
\arguments{
\item{x}{x}

\item{digits}{digits}

\item{...}{...}
}
\description{
S3 method for print for fs class
}
