% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesPriority.R
\name{prioritize.species}
\alias{prioritize.species}
\title{Subset species}
\usage{
prioritize.species(
  gene.list,
  orthologue.selection.directory,
  download.directory,
  species_priority_table = "reemplazos.txt"
)
}
\arguments{
\item{gene.list}{character; Gene name (file name) of the MSA}

\item{orthologue.selection.directory}{character; Output directory}

\item{download.directory}{character; Original directory where files are stored}

\item{species_priority_table}{data frame; Table specifying which species should be prioritized}
}
\value{
A fasta file with the selected species.
}
\description{
This function allows you to select only a few of the downloaded species to perform analysis.
}
\details{
The function takes a table of species available in your MSA ranked by "priority" and selects those to create a new fasta file.
}
\keyword{subset}
