% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunM0codeml.R
\name{run.m0.paml}
\alias{run.m0.paml}
\title{Run PAML M0 model}
\usage{
run.m0.paml(gene.list, input.directory, output.directory, software.directory)
}
\arguments{
\item{gene.list}{Gene name or file name}

\item{input.directory}{Directory where newick and phylip files are located}

\item{output.directory}{Directory where M0 optimized files will be written}

\item{software.directory}{Directory where PAML software is locateed}
}
\value{
A mlc file along with txt others files written by PAML. We mostly
 make use of the mlc file with the optimized tree.
}
\description{
Run PAML M0 model
}
