% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCodeml.R
\name{run.paml}
\alias{run.paml}
\title{Run Paml Codeml}
\usage{
run.paml(gene.list, software.directory, Ho.directory, Ha.directory)
}
\arguments{
\item{gene.list}{character; A gene list or file name}

\item{software.directory}{character; Directory where PAML Codeml executables are}

\item{Ho.directory}{character; Directory where the null hypothesis files will be written}

\item{Ha.directory}{character; Directory where the alternative hypothesis will be written}
}
\value{
Several txt files after the codeml run (such as the mlc)
}
\description{
Run Paml Codeml
}
