% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplealign2.r
\docType{data}
\name{sample.alignment.2}
\alias{sample.alignment.2}
\title{Sample alignment 2}
\format{
A text file with a sequence alignment in fasta format.
}
\source{
Sequences downloaded from the AAK1 gene in Ensembl
}
\usage{
sample.alignment.2
}
\description{
A dataset containing a sample alignment in fasta format.
}
\keyword{datasets}
