% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesPriority.R
\name{select.subset}
\alias{select.subset}
\title{Subset species}
\usage{
select.subset(
  gene.list,
  orthologue.selection.directory,
  download.directory,
  species_priority_table = "reemplazos.txt"
)
}
\arguments{
\item{gene.list}{Gene name (file name) of the MSA}

\item{orthologue.selection.directory}{Output directory}

\item{download.directory}{Original directory where files are stored}

\item{species_priority_table}{Table specifying which species should be prioritized}
}
\value{
A fasta file with the selected species.
}
\description{
This function allows you to select only a few of the downloaded species to perform analysis.
}
\details{
The function takes a table of species available in your MSA ranked by "priority" and selects those to create a new fasta file.
}
\examples{
\dontrun{genomic.locator(ensembl.protein.ID = "ENSP00000430945", original.number = 1269)}
}
\keyword{subset}
