% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectionr.r
\docType{package}
\name{selectionr}
\alias{selectionr}
\alias{selectionr-package}
\title{selectionr: Tools for working with MSA and Positive Selection Data}
\description{
Easily manage necessary files to run a positive selection test
}
