% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlowForegroundClassifier.R
\name{species.foreground.classifier}
\alias{species.foreground.classifier}
\title{Slow foreground classifier}
\usage{
species.foreground.classifier(
  gene.list,
  suffix = ".fasta",
  Taxon.id = "Mammalia",
  input.directory,
  output.directory
)
}
\arguments{
\item{gene.list}{character; Gene name or file name}

\item{suffix}{character; Suffix of the file name}

\item{Taxon.id}{numeric; Taxon to check whether foreground species belong}

\item{input.directory}{character; Directory where files are located}

\item{output.directory}{character; Directory where csv files will be written}
}
\value{
A csv file with species and class (either foreground or background)
}
\description{
Slow foreground classifier checks every single species at the alignment and
evaluates if they belong to a certain taxon or not. For several files, it is more
convenient to use the fast classifier.
}
