% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesArrange.R
\name{species.top}
\alias{species.top}
\title{Species at top of the alignment}
\usage{
species.top(
  gene.list,
  suffix = "_nuc_trimmed.fas",
  species.atop = "homo_sapiens",
  input.directory,
  output.directory
)
}
\arguments{
\item{gene.list}{The name of the gene}

\item{suffix}{The filename/type of the fasta file}

\item{species.atop}{The name of the species you want atop of the alignment (reference)}

\item{input.directory}{The input directory where the fasta file is stored}

\item{output.directory}{The output directory where you want the new fasta}
}
\value{
Returns a new fasta file with your desired species at top of the alignment
}
\description{
A function to make a species be a top of the alignment
}
\details{
Useful when some type of aligners rearrange species order. Works for DNA.
}
