% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryData.R
\name{summ.statistics.paml}
\alias{summ.statistics.paml}
\title{Summary statistics and positively selected sites}
\usage{
summ.statistics.paml(gene.list, Ho.directory, Ha.directory, summary.directory)
}
\arguments{
\item{gene.list}{Gene name (file name) of the MSA}

\item{Ho.directory}{Directory where PAML null hypothesis are located}

\item{Ha.directory}{Directory where PAML alternative hypothesis are located}

\item{summary.directory}{Directory summarized PAML files will be located}
}
\value{
A table with the selected sites for each gene
}
\description{
This function allows you to summarize several mlc files from PAML into a single table.
}
\keyword{summary}
