% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimalClean.R
\name{trimal.clean}
\alias{trimal.clean}
\title{Trimal cleaning}
\usage{
trimal.clean(
  gene.list,
  suffix = "_pranked.best.fas",
  alignment.directory,
  trim.output.directory,
  unaligned.directory,
  software.directory
)
}
\arguments{
\item{gene.list}{Gene name or file name}

\item{suffix}{Suffix for gene names}

\item{alignment.directory}{Directory where alignments are located}

\item{trim.output.directory}{Directory where trimmed sequences will be written}

\item{unaligned.directory}{Directory where unaligned sequences are located}

\item{software.directory}{Directory where Trimal executable is located}
}
\value{
A cleaned fasta file
}
\description{
Input aligned sequences as AA, trim them and then backtranslate to cds
}
