% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblDownload.R
\name{write.FASTA}
\alias{write.FASTA}
\title{Write a fasta file from Ensembl Orthologues}
\usage{
write.FASTA(JSON.object, out.file, gene.name, out.file2)
}
\arguments{
\item{JSON.object}{JSON object to be parsed}

\item{out.file}{Output file name}

\item{gene.name}{Gene name to be parsed}

\item{out.file2}{Not used parameter}
}
\value{
A txt file containing sequences and species names in fasta format.
}
\description{
This non-exported function creates a fasta file from Ensembl orthologues information.
}
\keyword{writeFasta}
