% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblDownload.R
\name{ensembl.orthologue.download}
\alias{ensembl.orthologue.download}
\title{Download Ensembl Orthologues}
\usage{
ensembl.orthologue.download(
  server = "https://jan2020.rest.ensembl.org",
  gene.name,
  type = "symbol",
  target_species = NULL,
  target_taxon,
  download.directory
)
}
\arguments{
\item{server}{Ensembl server (version)}

\item{gene.name}{A list of gene symbol names}

\item{type}{Type of gene name (either symbol or Ensembl Gene Name)}

\item{target_species}{Input target species to search for orthologues
for your gene symbol. Defaults to NULL. Only required when gene symbol is provided.}

\item{target_taxon}{The NCBI number of the species you'd like to download}

\item{download.directory}{The directory where you'd like the downloaded multiple sequences to be stored}
}
\value{
A txt file containing sequences and species names in fasta format.
}
\description{
This function allows you to download 1-to-1 orthologues from Ensembl, given a target taxon.
}
\examples{
\dontrun{ensembl.orthologue.download(gene.name = "RBFOX1", type = "gene.symbol",
target_species = "human",
 target_taxon = "9443", download.directory = ".")}
}
\keyword{download}
