% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicLocations.R
\name{genomic.locator}
\alias{genomic.locator}
\title{Get genomic location from Ensembl Protein}
\usage{
genomic.locator(
  server = "https://jul2019.rest.ensembl.org",
  ensembl.protein.ID,
  original.number
)
}
\arguments{
\item{server}{Ensembl version and server to query}

\item{ensembl.protein.ID}{A list of Ensembl protein IDs}

\item{original.number}{Aminoacid number you would like to get its genomic location}
}
\value{
A data frame with gene name, chr, start and end of original aminoacidic position.
}
\description{
This function allows you to get the geneomic location from aminoacid sites obtained from Ensembl proteins.
}
\examples{
\dontrun{genomic.locator(ensembl.protein.ID = "ENSP00000430945", original.number = 1269)}
}
\keyword{genomic}
\keyword{location}
