% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginalPosition.R
\name{positions.provean}
\alias{positions.provean}
\title{Get original positions in alignment before masking}
\usage{
positions.provean(
  gene.name,
  input.directory,
  suffix,
  output.directory,
  classifier.directory,
  beb.sites.directory,
  target_taxon = 9443
)
}
\arguments{
\item{gene.name}{character; Gene name (file name). Gene symbols from human only.}

\item{input.directory}{character; Directory where the files are}

\item{suffix}{character; Suffix for this file (".fasta", "_trimmed.fas", etc)}

\item{output.directory}{character; Directory where the files will be written}

\item{classifier.directory}{character; Directory where the species were classified
as foreground and background}

\item{beb.sites.directory}{character; Directory where the mlc with
the BEB sites are located}

\item{target_taxon}{numeric; Reference species taxon number id (NCBI)}
}
\value{
A table with Ensembl protein ID, original position and aminoacid and outgroup aminoacid
}
\description{
Use BEB sites obtained from a masked (trimmed) alignment and find the positions
 in the original alignment. Useful when the starting aslignment was the CDS from
 Ensembl. Therefore, sites are the original protein positions from the Ensembl protein
 and are easily traceable. It also reports the most frequent aminoacid from the background
 sequences (excluding gaps).
}
