% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrankAligner.R
\name{prank.align}
\alias{prank.align}
\title{A PRANK aligner wrapper}
\usage{
prank.align(
  gene.list,
  input.file.directory,
  output.file.directory,
  software.directory
)
}
\arguments{
\item{gene.list}{List of gene names (file names) to align}

\item{input.file.directory}{Directory where the fasta files with sequences
are located}

\item{output.file.directory}{Directory where the fasta files with
MSA will be written}

\item{software.directory}{Directory where your prank executable is located}
}
\value{
An aligned fasta file
}
\description{
A function to align sequences with default prank settings.
}
