% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrankAligner.R
\name{prank.align}
\alias{prank.align}
\title{A PRANK aligner wrapper}
\usage{
prank.align(
  gene.name,
  input.directory,
  output.file.directory,
  software.directory
)
}
\arguments{
\item{gene.name}{character; List of gene names (file names) to align}

\item{input.directory}{character; Directory where the fasta files with sequences
are located}

\item{output.file.directory}{character; Directory where the fasta files with
MSA will be written}

\item{software.directory}{character; Directory where your prank executable is located}
}
\value{
An aligned fasta file
}
\description{
A function to align sequences with default prank settings.
}
\examples{
\dontrun{prank.align(sample.alignment.2, input.directory = "~/data/",
output.directory = "~/data/", software.directory = "/software")}
}
