% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrequalClean.R
\name{prequal.clean}
\alias{prequal.clean}
\title{Prequal cleaner wrapper}
\usage{
prequal.clean(
  gene.list,
  suffix = ".fasta",
  unaligned.directory,
  trim.output.directory,
  software.directory
)
}
\arguments{
\item{gene.list}{A list of gene names or file names.}

\item{suffix}{Suffix for the file names (".fasta", etc)}

\item{unaligned.directory}{Directory where unaligned sequences are stored}

\item{trim.output.directory}{Directory where cleaned sequences will be written}

\item{software.directory}{Directory where prequal is located}
}
\value{
A fasta file with trimmed sequences
}
\description{
Prequal cleaner wrapper
}
